refactor(backend): split llm_client.py into app/llm/ package with facade (#582)
app/utils/llm_client.py (1635 LOC) aufgeteilt:

- app/llm/client.py    — LLMClient, Tool-Calling-Helpers,
                         build_client_from_profile (git mv, History erhalten)
- app/llm/context.py   — _MODEL_CONTEXT_HEURISTICS, heuristic_num_ctx_for_model,
                         _warn_legacy_fallback_once, _resolve_num_ctx
- app/llm/json_mode.py — Schema-Aufbereitung (strict/Ollama-format),
                         Env-Flags, should_disable_openai_json_mode,
                         _strip_llm_json_envelope, _try_repair_truncated_json
- app/llm/providers/, errors.py, retry.py — bereits in #591/#590 gelandet

app/utils/llm_client.py bleibt als Fassade: statische Re-Exports fuer
Type-Checker plus sys.modules-Alias auf app.llm.client — damit treffen
bestehende Monkeypatches/Mocks (z. B. "app.utils.llm_client.OpenAI",
"….llm_call_with_retry", "…._read_active_config_safely") weiterhin das
Implementations-Modul. Alle Bestands-Importe und die komplette Testsuite
laufen unveraendert (2832 passed).

context.py/json_mode.py loggen unter demselben Logger-Namen
("agora.llm_client"), damit patch.object(llm_client.logger, …)-Tests
die Warnings weiter sehen.

mypy: app.llm.client uebernimmt die Legacy-Baseline (ignore_errors)
aus app.utils.*; die neuen app.llm-Module laufen ohne Override —
uv run mypy app: Success (203 Dateien). Public-Signaturen unveraendert,
reines Restrukturieren.

PLAN.md: Modulbaum + Phase-1-Verweise auf app/llm/ aktualisiert.

Teil von Milestone M3 (#636).
